requirements.txt
# ====================================
# AI Stock Dashboard Requirements
# Version: 2.0
# Updated: October 5, 2025
# ====================================

# Core Dependencies
streamlit>=1.28.0
yfinance>=0.2.18
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.15.0
scikit-learn>=1.3.0

# News & Web Scraping
feedparser>=6.0.10           # RSS feed parsing
beautifulsoup4>=4.12.0       # Web scraping
requests>=2.31.0             # HTTP requests
lxml>=4.9.0                  # XML/HTML parsing
# selenium>=4.15.0           # Optional: Dynamic scraping

# Sentiment Analysis
vaderSentiment>=3.3.2        # Financial sentiment (VADER)
textblob>=0.17.1             # NLP analysis
# transformers>=4.35.0       # Optional: FinBERT (large dependency)
# torch>=2.1.0               # Optional: PyTorch for FinBERT

# Alerts & Notifications
plyer>=2.1.0                 # Desktop notifications (cross-platform)
python-telegram-bot>=20.7    # Telegram bot API
pygame>=2.5.0                # Audio alerts

# Social Media APIs
praw>=7.7.1                  # Reddit API
# tweepy>=4.14.0             # Optional: Twitter API

# Configuration & Utilities
pyyaml>=6.0.1                # YAML config parsing
python-dotenv>=1.0.0         # Environment variables
schedule>=1.2.0              # Task scheduling
tabulate>=0.9.0              # Pretty tables for CLI
pytz>=2023.3                 # Timezone handling (REQUIRED)

# Data Validation & Quality
cerberus>=1.3.5              # Schema validation

# Optional: Advanced Features
# alpaca-trade-api>=3.0.0    # Alpaca broker integration
# ib-insync>=0.9.86          # Interactive Brokers
# ccxt>=4.1.0                # Cryptocurrency exchanges